+++ app/addactivity.tsx
@@ -12,29 +12,32 @@ import {
   View,
   TouchableOpacity,
   Alert,
-  SafeAreaView, 
+  SafeAreaView,
 } from "react-native";
 import DateTimePicker from "@react-native-community/datetimepicker";
-import { formattedDate } from "../utils/formattedDate";
-import { useDate } from "../providers/DateProvider";
 
-interface FormData {
+import { useDate } from '../providers/DateProvider';
+import { prettyDate } from '../utils/prettyDate';
+import useActivity from '../hooks/useActivity';
+
+type FormData = {
   label: string;
   description: string;
   startTime: Date;
   endTime: Date;
-}
+};
 
 const AddActivity = () => {
   const router = useRouter();
   const { selectedDate } = useDate();
   const [isStartTimePickerVisible, setStartTimePickerVisible] = useState(false);
   const [isEndTimePickerVisible, setEndTimePickerVisible] = useState(false);
+  const { useCreateActivity } = useActivity({ date: selectedDate });
   const [formData, setFormData] = useState<FormData>({
-    label: "",
-    description: "",
+    label: '',
+    description: '',
     startTime: new Date(),
-    endTime: new Date(new Date().setHours(23, 59)),
+    endTime: new Date(),
   });
 
   const handleInputChange = (field: keyof FormData, value: string | Date) => {
@@ -46,16 +49,15 @@ const AddActivity = () => {
 
   const formatTime = (date: Date) => {
     // Format the time as HH:MM
-    return date.toLocaleTimeString("en-GB", {
+    return date.toLocaleTimeString('it-IT', {
       hour: "2-digit",
       minute: "2-digit",
     });
   };
 
-  const handleSubmit = () => {
+  const handleSubmit = async () => {
     const { label, description, startTime, endTime } = formData;
 
-    const formattedDate = selectedDate.toDateString();
     const formattedStartTime = formatTime(startTime);
     const formattedEndTime = formatTime(endTime);
 
@@ -65,102 +67,111 @@ const AddActivity = () => {
       return;
     }
 
-    console.log(
-      `Adding activity on ${formattedDate} with label ${label}, description ${description}, start time ${formattedStartTime}, and end time ${formattedEndTime}`
-    );
+    await useCreateActivity.mutateAsync({
+      citizenId: 1,
+      data: {
+        activityId: -1,
+        name: label,
+        description,
+        startTime: formattedStartTime,
+        endTime: formattedEndTime,
+        date: selectedDate.toISOString().split('T')[0],
+        isCompleted: false,
+      },
+    });
     router.back();
   };
 
   return (
-    <SafeAreaView style={styles.container}>
-      <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
-        <KeyboardAvoidingView
-          style={styles.container}
-          behavior={Platform.OS === "ios" ? "padding" : undefined} //Android's built-in handling should suffice
-          keyboardVerticalOffset={80}
-        >
-          <ScrollView contentContainerStyle={{ flexGrow: 1 }}>
-            <Text style={styles.headerText}>
-              Opret en aktivitet til {formattedDate(selectedDate)}
-            </Text>
-  
-            <TextInput
-              style={styles.input}
-              placeholder="Navn"
-              value={formData.label}
-              onChangeText={(text) => handleInputChange("label", text)}
-              returnKeyType="done"
-            />
-  
-            <TextInput
-              style={styles.description}
-              placeholder="Beskrivelse"
-              value={formData.description}
-              onChangeText={(text) => handleInputChange("description", text)}
-              multiline
-              returnKeyType="done"
-            />
-  
-            <View style={styles.pickerContainer}>
-              <Text style={styles.header}>Vælg start tid</Text>
-              <TouchableOpacity onPress={() => setStartTimePickerVisible(true)}>
-                <Text>{formatTime(formData.startTime)}</Text>
-              </TouchableOpacity>
-            </View>
+      <SafeAreaView style={styles.container}>
+        <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
+          <KeyboardAvoidingView
+              style={styles.container}
+              behavior={Platform.OS === "ios" ? "padding" : undefined} //Android's built-in handling should suffice
+              keyboardVerticalOffset={80}
+          >
+            <ScrollView contentContainerStyle={{ flexGrow: 1 }}>
+              <Text style={styles.headerText}>
+                Opret en aktivitet til {prettyDate(selectedDate)}
+              </Text>
 
-            {/*When StartTimePickerVisible is set to true it shows the DateTimePicker for EndTime*/}
-            {isStartTimePickerVisible && (
-                <DateTimePicker
-                    mode="time"
-                    value={formData.startTime}
-                    is24Hour={true}
-                    display="spinner"
-                    minuteInterval={5}
-                    onChange={(_event, selectedTime) => {
-                      setStartTimePickerVisible(false);
-                      if (selectedTime) {
-                        handleInputChange("startTime", selectedTime);
-                      }
-                    }}
-                    style={styles.timePicker}
-                />
-            )}
-  
-            <View style={styles.pickerContainer}>
-              <Text style={styles.header}>Vælg slut tid</Text>
-              <TouchableOpacity onPress={() => setEndTimePickerVisible(true)}>
-                <Text>{formatTime(formData.endTime)}</Text>
-              </TouchableOpacity>
-            </View>
+              <TextInput
+                  style={styles.input}
+                  placeholder="Navn"
+                  value={formData.label}
+                  onChangeText={(text) => handleInputChange("label", text)}
+                  returnKeyType="done"
+              />
 
-            {/*When EndTimePickerVisible is set to true it shows the DateTimePicker for EndTime*/}
-            {isEndTimePickerVisible && (
-                <DateTimePicker
-                    mode="time"
-                    value={formData.endTime}
-                    is24Hour={true}
-                    display="spinner"
-                    minuteInterval={5}
-                    onChange={(_event, selectedTime) => {
-                      setEndTimePickerVisible(false); 
-                      if (selectedTime) {
-                        handleInputChange("endTime", selectedTime);
-                      }
-                    }}
-                    style={styles.timePicker}
-                />
-            )}
-  
-            <TouchableOpacity
-              style={[styles.button, styles.addButton]}
-              onPress={handleSubmit}
-            >
-              <Text style={styles.buttonText}>Tilføj</Text>
-            </TouchableOpacity>
-          </ScrollView>
-        </KeyboardAvoidingView>
-      </TouchableWithoutFeedback>
-    </SafeAreaView>
+              <TextInput
+                  style={styles.description}
+                  placeholder="Beskrivelse"
+                  value={formData.description}
+                  onChangeText={(text) => handleInputChange("description", text)}
+                  multiline
+                  returnKeyType="done"
+              />
+
+              <View style={styles.pickerContainer}>
+                <Text style={styles.header}>Vælg start tid</Text>
+                <TouchableOpacity onPress={() => setStartTimePickerVisible(true)}>
+                  <Text>{formatTime(formData.startTime)}</Text>
+                </TouchableOpacity>
+              </View>
+
+              {/*When StartTimePickerVisible is set to true it shows the DateTimePicker for EndTime*/}
+              {isStartTimePickerVisible && (
+                  <DateTimePicker
+                      mode="time"
+                      value={formData.startTime}
+                      is24Hour={true}
+                      display="spinner"
+                      minuteInterval={5}
+                      onChange={(_event, selectedTime) => {
+                        setStartTimePickerVisible(false);
+                        if (selectedTime) {
+                          handleInputChange("startTime", selectedTime);
+                        }
+                      }}
+                      style={styles.timePicker}
+                  />
+              )}
+
+              <View style={styles.pickerContainer}>
+                <Text style={styles.header}>Vælg slut tid</Text>
+                <TouchableOpacity onPress={() => setEndTimePickerVisible(true)}>
+                  <Text>{formatTime(formData.endTime)}</Text>
+                </TouchableOpacity>
+              </View>
+
+              {/*When EndTimePickerVisible is set to true it shows the DateTimePicker for EndTime*/}
+              {isEndTimePickerVisible && (
+                  <DateTimePicker
+                      mode="time"
+                      value={formData.endTime}
+                      is24Hour={true}
+                      display="spinner"
+                      minuteInterval={5}
+                      onChange={(_event, selectedTime) => {
+                        setEndTimePickerVisible(false);
+                        if (selectedTime) {
+                          handleInputChange("endTime", selectedTime);
+                        }
+                      }}
+                      style={styles.timePicker}
+                  />
+              )}
+
+              <TouchableOpacity
+                  style={[styles.button, styles.addButton]}
+                  onPress={handleSubmit}
+              >
+                <Text style={styles.buttonText}>Tilføj</Text>
+              </TouchableOpacity>
+            </ScrollView>
+          </KeyboardAvoidingView>
+        </TouchableWithoutFeedback>
+      </SafeAreaView>
   );
 };
